done service provider and mapping

+++ apps/backend/src/controllers/service.controller.js
@@ -1,5 +1,6 @@
 import {
-  ServiceProviderService,
+  MappingService,
+  ProviderService,
   ServiceService,
 } from "../services/service.service.js";
 import { ApiError } from "../utils/ApiError.js";
@@ -11,21 +12,15 @@ class ServiceProviderController {
   static create = asyncHandler(async (req, res) => {
     const { type } = req.body;
 
-    if (!type) throw ApiError.badRequest("Type is required");
-
     let result;
 
-    if (type === "provider") {
-      result = await ServiceProviderService.create(req.body);
-    } else if (type === "service") {
-      result = await ServiceService.create(req.body);
-    } else {
-      throw ApiError.badRequest("Invalid type. Use provider or service");
-    }
+    if (type === "service") result = await ServiceService.create(req.body);
+    else if (type === "provider")
+      result = await ProviderService.create(req.body);
+    else if (type === "mapping") result = await MappingService.create(req.body);
+    else throw ApiError.badRequest("Invalid type");
 
-    return res
-      .status(201)
-      .json(ApiResponse.success(result, `${type} created successfully`));
+    res.status(201).json(ApiResponse.success(result));
   });
 
   // UPDATE
@@ -38,10 +33,12 @@ class ServiceProviderController {
 
     let result;
 
-    if (type === "provider") {
-      result = await ServiceProviderService.update(id, req.body);
-    } else if (type === "service") {
+    if (type === "service") {
       result = await ServiceService.update(id, req.body);
+    } else if (type === "provider") {
+      result = await ProviderService.update(id, req.body);
+    } else if (type === "mapping") {
+      result = await MappingService.update(id, req.body);
     } else {
       throw ApiError.badRequest("Invalid type");
     }
@@ -53,27 +50,34 @@ class ServiceProviderController {
 
   // GET ALL (Filter + Search + Pagination)
   static getAll = asyncHandler(async (req, res) => {
-    const { type, page, limit, search, isActive, providerId } = req.query;
+    const { type, page, limit, search, isActive } = req.body;
+
+    console.log(type);
+    console.log(isActive);
 
     if (!type) throw ApiError.badRequest("Type is required");
 
+    const pagination = {
+      page: Number(page) || 1,
+      limit: Number(limit) || 10,
+    };
+
     let result;
 
-    if (type === "provider") {
-      result = await ServiceProviderService.getAll({
-        page: Number(page) || 1,
-        limit: Number(limit) || 10,
-        search,
-        isActive: isActive !== undefined ? isActive === "true" : undefined,
-      });
-    } else if (type === "service") {
+    if (type === "service") {
       result = await ServiceService.getAll({
-        page: Number(page) || 1,
-        limit: Number(limit) || 10,
+        ...pagination,
         search,
-        providerId,
-        isActive: isActive !== undefined ? isActive === "true" : undefined,
+        isActive: isActive !== undefined ? isActive === true : undefined,
       });
+    } else if (type === "provider") {
+      result = await ProviderService.getAll({
+        ...pagination,
+        search,
+        isActive: isActive !== undefined ? isActive === true : undefined,
+      });
+    } else if (type === "mapping") {
+      result = await MappingService.getAll(pagination);
     } else {
       throw ApiError.badRequest("Invalid type");
     }
@@ -87,13 +91,16 @@ class ServiceProviderController {
     const { id } = req.params;
 
     if (!type) throw ApiError.badRequest("Type is required");
+    if (!id) throw ApiError.badRequest("ID is required");
 
     let result;
 
-    if (type === "provider") {
-      result = await ServiceProviderService.delete(id);
-    } else if (type === "service") {
+    if (type === "service") {
       result = await ServiceService.delete(id);
+    } else if (type === "provider") {
+      result = await ProviderService.delete(id);
+    } else if (type === "mapping") {
+      result = await MappingService.delete(id);
     } else {
       throw ApiError.badRequest("Invalid type");
     }
